requirements.txt
asgiref
black
django
# CMS Plugins
django-cms
django-filer
djangocms-text-ckeditor
djangocms-link
djangocms-file
djangocms-picture
djangocms-video
djangocms-googlemap
djangocms-snippet
djangocms-style
django-cors-headers
django-ckeditor
django-debug-toolbar
django-extensions
django-heroku
django-markdownx
django-redis
djangorestframework
django-sass-processor
gunicorn
importlib-metadata
libsass
Markdown
Pillow
pytz
rcssmin
redis
rjsmin
six
sqlparse
typing-extensions
uwsgi
zipp
pre-commit
